// src-web/types/ContentConfig.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ClientConfig } from './ClientConfig'

export type ContentConfig = {
  method: string
  url: string
  headers: Array<[string, string]>
  data: Array<number> | null
  client: ClientConfig
}